Move messages.ts to device-api
Messages are all related to Supervisor API
